Resolved uint/int issues in clock publisher due to ROS1
1. Changed ROS connection to ROS 1 from ROS 2
2. Resolved uint/int issues in timestam and time message by typecasting.

+++ Nav2SLAMExampleProject/Assets/Scripts/RosInterfaces/LaserScanSensor.cs
@@ -98,7 +98,7 @@ namespace Unity.Robotics.Nav2SlamExample
                     frame_id = m_FrameId,
                     stamp = new TimeMsg
                     {
-                        sec = timestamp.Seconds,
+                        sec = (uint)timestamp.Seconds,
                         nanosec = timestamp.NanoSeconds,
                     }
                 },
